# src/main/resources/openapi.yaml
openapi: 3.1.0
info:
  title: BlueBubbles ChatGPT Agent OpenAPI spec
  description: Agent that bridges ChatGPT and iMessage
  version: 0.0.1
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/v1/admin/list.subscriptions:
    get:
      operationId: adminListSubscriptions
      tags:
        - admin
      summary: Lists subscription state for admins
      description: Returns payment subscription stats and recent subscription rows.
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 100
      responses:
        '200':
          description: Subscription admin list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminSubscriptionListResponse'
        '401':
          description: Authentication required
        '403':
          description: Admin role required
  /api/v1/admin/sync.subscription:
    post:
      operationId: adminSyncSubscription
      tags:
        - admin
      summary: Syncs one provider subscription
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminSubscriptionActionRequest'
      responses:
        '200':
          description: Subscription synced
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminSubscriptionActionResponse'
        '400':
          description: Invalid request
        '401':
          description: Authentication required
        '403':
          description: Admin role required
        '404':
          description: Subscription not found
  /api/v1/admin/suspend.subscription:
    post:
      operationId: adminSuspendSubscription
      tags:
        - admin
      summary: Suspends one provider subscription
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminSubscriptionActionRequest'
      responses:
        '200':
          description: Subscription suspended
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminSubscriptionActionResponse'
        '400':
          description: Invalid request
        '401':
          description: Authentication required
        '403':
          description: Admin role required
        '404':
          description: Subscription not found
  /api/v1/admin/unsuspend.subscription:
    post:
      operationId: adminUnsuspendSubscription
      tags:
        - admin
      summary: Unsuspends one provider subscription
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminSubscriptionActionRequest'
      responses:
        '200':
          description: Subscription unsuspended
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminSubscriptionActionResponse'
        '400':
          description: Invalid request
        '401':
          description: Authentication required
        '403':
          description: Admin role required
        '404':
          description: Subscription not found
  /api/v1/admin/grantPremium.subscription:
    post:
      operationId: adminGrantPremium
      tags:
        - admin
      summary: Grants manual premium access
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminPremiumGrantRequest'
      responses:
        '200':
          description: Manual premium granted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminSubscriptionActionResponse'
        '400':
          description: Invalid request
        '401':
          description: Authentication required
        '403':
          description: Admin role required
        '404':
          description: Account not found
  /api/v1/admin/revokePremium.subscription:
    post:
      operationId: adminRevokePremium
      tags:
        - admin
      summary: Revokes manual premium access
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminPremiumGrantRequest'
      responses:
        '200':
          description: Manual premium revoked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminSubscriptionActionResponse'
        '400':
          description: Invalid request
        '401':
          description: Authentication required
        '403':
          description: Admin role required
        '404':
          description: Account not found
  /api/v1/admin/get.statistics:
    get:
      operationId: adminGetStatistics
      tags:
        - admin
      summary: Gets recent agent statistics
      description: Returns message, model, tool, and active-user statistics for an admin-selected time period.
      security:
        - bearerAuth: []
      parameters:
        - name: from
          in: query
          description: Inclusive start of the statistics period. Defaults to 24 hours before now.
          required: false
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          description: Exclusive end of the statistics period. Defaults to now.
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Admin statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminStatsResponse'
        '400':
          description: Invalid time period
        '401':
          description: Authentication required
        '403':
          description: Admin role required
  /api/v1/admin/get.rate_limit_usage:
    get:
      operationId: adminGetRateLimitUsage
      tags:
        - admin
      summary: Gets current rate-limit usage
      description: Returns current usage for app rate limits, including daily assistant response usage by account.
      security:
        - bearerAuth: []
      parameters:
        - name: limit_key
          in: query
          description: Rate limit key to inspect. Defaults to daily assistant response usage.
          required: false
          schema:
            type: string
        - name: limit
          in: query
          description: Maximum number of usage rows to return.
          required: false
          schema:
            type: integer
            format: int32
            default: 50
            minimum: 1
            maximum: 500
      responses:
        '200':
          description: Admin rate-limit usage
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminRateLimitUsageResponse'
        '401':
          description: Authentication required
        '403':
          description: Admin role required
  /api/v1/admin/list.feedback:
    get:
      operationId: adminListFeedback
      tags:
        - admin
      summary: Lists agent feedback
      description: Returns user feedback captured by the agent feedback tool as an admin inbox.
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          description: Feedback read state to list. Defaults to unread.
          required: false
          schema:
            type: string
            enum:
              - all
              - unread
              - read
            default: unread
        - name: limit
          in: query
          description: Maximum number of feedback items to return.
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 250
            default: 100
      responses:
        '200':
          description: Feedback inbox items
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminFeedbackListResponse'
        '401':
          description: Authentication required
        '403':
          description: Admin role required
  /api/v1/admin/markRead.feedback:
    post:
      operationId: adminMarkFeedbackRead
      tags:
        - admin
      summary: Marks feedback as read
      description: Marks a feedback inbox item as read.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminFeedbackStatusUpdateRequest'
      responses:
        '200':
          description: Feedback item after the update
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminFeedbackItem'
        '400':
          description: Missing feedback id
        '401':
          description: Authentication required
        '403':
          description: Admin role required
        '404':
          description: Feedback item not found
  /api/v1/admin/markUnread.feedback:
    post:
      operationId: adminMarkFeedbackUnread
      tags:
        - admin
      summary: Marks feedback as unread
      description: Marks a feedback inbox item as unread.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminFeedbackStatusUpdateRequest'
      responses:
        '200':
          description: Feedback item after the update
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminFeedbackItem'
        '400':
          description: Missing feedback id
        '401':
          description: Authentication required
        '403':
          description: Admin role required
        '404':
          description: Feedback item not found
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AdminSubscriptionStats:
      type: object
      required:
        - total_subscriptions
        - active_subscriptions
        - past_due_subscriptions
        - inactive_subscriptions
        - monthly_recurring_amount
        - currency
      properties:
        total_subscriptions:
          type: integer
          format: int64
        active_subscriptions:
          type: integer
          format: int64
        past_due_subscriptions:
          type: integer
          format: int64
        inactive_subscriptions:
          type: integer
          format: int64
        monthly_recurring_amount:
          type: string
        currency:
          type: string
    AdminSubscriptionItem:
      type: object
      required:
        - subscription_id
        - account_id
        - provider
        - plan_key
        - status
        - updated_at
      properties:
        subscription_id:
          type: string
        account_id:
          type: string
        account_bucket:
          type: string
          nullable: true
        provider:
          type: string
        plan_key:
          type: string
        status:
          type: string
        provider_status:
          type: string
          nullable: true
        provider_subscription_id:
          type: string
          nullable: true
        provider_customer_id:
          type: string
          nullable: true
        current_period_end:
          type: string
          format: date-time
          nullable: true
        grace_period_end:
          type: string
          format: date-time
          nullable: true
        management_url:
          type: string
          nullable: true
        last_synced_at:
          type: string
          format: date-time
          nullable: true
        updated_at:
          type: string
          format: date-time
    AdminSubscriptionListResponse:
      type: object
      required:
        - generated_at
        - stats
        - subscriptions
      properties:
        generated_at:
          type: string
          format: date-time
        stats:
          $ref: '#/components/schemas/AdminSubscriptionStats'
        subscriptions:
          type: array
          items:
            $ref: '#/components/schemas/AdminSubscriptionItem'
    AdminSubscriptionActionRequest:
      type: object
      required:
        - subscription_id
      properties:
        subscription_id:
          type: string
        reason:
          type: string
          nullable: true
    AdminPremiumGrantRequest:
      type: object
      required:
        - account_id
      properties:
        account_id:
          type: string
    AdminSubscriptionActionResponse:
      type: object
      required:
        - status
      properties:
        status:
          type: string
        account_id:
          type: string
          nullable: true
        subscription_id:
          type: string
          nullable: true
        subscription:
          $ref: '#/components/schemas/AdminSubscriptionItem'
    AdminStatsPeriod:
      type: object
      required:
        - from
        - to
        - bucket_size
      properties:
        from:
          type: string
          format: date-time
        to:
          type: string
          format: date-time
        bucket_size:
          type: string
          enum:
            - hour
            - day
    AdminModelStats:
      type: object
      required:
        - model_key
        - model_label
        - responses_model
        - is_premium
        - message_count
        - active_users
        - percentage
      properties:
        model_key:
          type: string
        model_label:
          type: string
        responses_model:
          type: string
        is_premium:
          type: boolean
        message_count:
          type: integer
          format: int64
        active_users:
          type: integer
          format: int64
        percentage:
          type: number
          format: double
    AdminBucketModelStats:
      type: object
      required:
        - model_key
        - model_label
        - message_count
      properties:
        model_key:
          type: string
        model_label:
          type: string
        message_count:
          type: integer
          format: int64
    AdminSenderStats:
      type: object
      required:
        - account_key_hash
        - account_bucket
        - message_count
        - percentage
        - last_seen_at
        - models
      properties:
        account_key_hash:
          type: string
          description: SHA-256 hash of the sender/account bucket used for metrics grouping.
        account_bucket:
          type: string
          description: Short stable display bucket derived from the account hash.
        message_count:
          type: integer
          format: int64
        percentage:
          type: number
          format: double
        last_seen_at:
          type: string
          format: date-time
        models:
          type: array
          items:
            $ref: '#/components/schemas/AdminBucketModelStats'
    AdminToolStats:
      type: object
      required:
        - tool_name
        - tool_category
        - call_count
        - successful_calls
        - failed_calls
        - active_users
        - success_rate
        - average_duration_ms
        - last_used_at
        - percentage
      properties:
        tool_name:
          type: string
        tool_category:
          type: string
          description: Coarse tool grouping such as bluebubbles, google_calendar, coder, coder_mcp, memory, website, or other.
        call_count:
          type: integer
          format: int64
        successful_calls:
          type: integer
          format: int64
        failed_calls:
          type: integer
          format: int64
        active_users:
          type: integer
          format: int64
        success_rate:
          type: number
          format: double
        average_duration_ms:
          type: number
          format: double
        last_used_at:
          type: string
          format: date-time
        percentage:
          type: number
          format: double
    AdminToolAccountTypeStats:
      type: object
      required:
        - is_premium
        - call_count
        - successful_calls
        - failed_calls
        - active_users
        - success_rate
        - percentage
      properties:
        is_premium:
          type: boolean
        call_count:
          type: integer
          format: int64
        successful_calls:
          type: integer
          format: int64
        failed_calls:
          type: integer
          format: int64
        active_users:
          type: integer
          format: int64
        success_rate:
          type: number
          format: double
        percentage:
          type: number
          format: double
    AdminStatsBucket:
      type: object
      required:
        - bucket_start
        - bucket_end
        - message_count
        - active_users
        - models
      properties:
        bucket_start:
          type: string
          format: date-time
        bucket_end:
          type: string
          format: date-time
        message_count:
          type: integer
          format: int64
        active_users:
          type: integer
          format: int64
        models:
          type: array
          items:
            $ref: '#/components/schemas/AdminBucketModelStats'
    AdminStatsResponse:
      type: object
      required:
        - period
        - total_messages
        - active_users
        - average_messages_per_user
        - total_tool_calls
        - successful_tool_calls
        - failed_tool_calls
        - tool_success_rate
        - models
        - senders
        - timeline
        - tools
        - tool_account_types
      properties:
        period:
          $ref: '#/components/schemas/AdminStatsPeriod'
        total_messages:
          type: integer
          format: int64
        active_users:
          type: integer
          format: int64
        average_messages_per_user:
          type: number
          format: double
        total_tool_calls:
          type: integer
          format: int64
        successful_tool_calls:
          type: integer
          format: int64
        failed_tool_calls:
          type: integer
          format: int64
        tool_success_rate:
          type: number
          format: double
        models:
          type: array
          items:
            $ref: '#/components/schemas/AdminModelStats'
        senders:
          type: array
          description: Top sender/account buckets by message volume for the selected period.
          items:
            $ref: '#/components/schemas/AdminSenderStats'
        timeline:
          type: array
          items:
            $ref: '#/components/schemas/AdminStatsBucket'
        tools:
          type: array
          description: Tool usage grouped by tool name for the selected period.
          items:
            $ref: '#/components/schemas/AdminToolStats'
        tool_account_types:
          type: array
          description: Tool usage grouped by account type for the selected period.
          items:
            $ref: '#/components/schemas/AdminToolAccountTypeStats'
    AdminRateLimitUsage:
      type: object
      required:
        - limit_key
        - limit_label
        - scope_type
        - scope_key
        - account_id
        - account_bucket
        - is_premium
        - used
        - limit
        - remaining
        - percentage
        - exhausted
        - window_start
        - window_end
        - updated_at
      properties:
        limit_key:
          type: string
          description: Stable app limit key, such as message_responses_per_day.
        limit_label:
          type: string
        scope_type:
          type: string
          description: Generic limit scope type, such as account.
        scope_key:
          type: string
          description: Scope identifier. For account-scoped limits this is the canonical account id.
        account_id:
          type: string
          nullable: true
          description: Canonical agent account id when the scope is an account.
        account_bucket:
          type: string
          nullable: true
          description: Short stable display bucket derived from the account id.
        is_premium:
          type: boolean
        used:
          type: integer
          format: int64
        limit:
          type: integer
          format: int64
        remaining:
          type: integer
          format: int64
        percentage:
          type: number
          format: double
        exhausted:
          type: boolean
        window_start:
          type: string
          format: date-time
        window_end:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    AdminRateLimitUsageResponse:
      type: object
      required:
        - generated_at
        - limit_key
        - usages
      properties:
        generated_at:
          type: string
          format: date-time
        limit_key:
          type: string
        usages:
          type: array
          items:
            $ref: '#/components/schemas/AdminRateLimitUsage'
    AdminFeedbackItem:
      type: object
      required:
        - feedback_id
        - account_id
        - account_bucket
        - submitted_at
        - feedback_text
        - category
        - transport
        - read_status
      properties:
        feedback_id:
          type: string
        account_id:
          type: string
          description: Canonical agent account id that submitted the feedback.
        account_bucket:
          type: string
          description: Short display bucket derived from the canonical account id.
        submitted_at:
          type: string
          format: date-time
        feedback_text:
          type: string
        category:
          type: string
        transport:
          type: string
        sender:
          type: string
          nullable: true
        chat_guid:
          type: string
          nullable: true
        message_guid:
          type: string
          nullable: true
        read_status:
          type: string
          enum:
            - unread
            - read
        read_at:
          type: string
          format: date-time
          nullable: true
    AdminFeedbackListResponse:
      type: object
      required:
        - status
        - items
        - unread_count
        - read_count
        - total_count
      properties:
        status:
          type: string
          enum:
            - all
            - unread
            - read
        items:
          type: array
          items:
            $ref: '#/components/schemas/AdminFeedbackItem'
        unread_count:
          type: integer
          format: int64
        read_count:
          type: integer
          format: int64
        total_count:
          type: integer
          format: int64
    AdminFeedbackStatusUpdateRequest:
      type: object
      required:
        - feedback_id
      properties:
        feedback_id:
          type: string
...